+++ main.js
@@ -21,9 +21,25 @@ Object.defineProperty(exports, "Thing", {
     return _Thing.Thing;
   }
 });
+Object.defineProperty(exports, "WebPage", {
+  enumerable: true,
+  get: function get() {
+    return _WebPage.WebPage;
+  }
+});
+Object.defineProperty(exports, "WebSite", {
+  enumerable: true,
+  get: function get() {
+    return _WebSite.WebSite;
+  }
+});
 
 var _Organization = require("./dist/Organization/Organization");
 
 var _Person = require("./dist/Person/Person");
 
 var _Thing = require("./dist/Thing/Thing");
+
+var _WebPage = require("./dist/WebPage/WebPage");
+
+var _WebSite = require("./dist/WebSite/WebSite");