+++ src/Types/PersonOrOrg.tsx
@@ -1,0 +1,6 @@
+import { PersonProps } from '../Person/Person'
+import { OrgProps } from '../Organization/Organization'
+
+type PersonOrOrganization = PersonProps | OrgProps
+
+export default PersonOrOrganization